refactor(runtime): split methods.rs into cohesive submodules (§7-8)
Split the 3493-line runtime/methods.rs (over the 500-line limit) into
two cohesive modules under src/runtime/. The impl block contains a
single indivisible 3183-line call_method_with_values method — this is
unavoidable; it cannot be split at a function boundary and is kept
alone in its own file per project convention.

- methods.rs (311)              - multidim helpers, allomorph helpers (free fns)
- methods_call_dispatch.rs (3197) - call_method_with_values (single indivisible method)

Pure relocation. Nine free functions bumped from private/pub(super) to
pub(crate) so they are accessible from the sibling file: multidim_at_pos,
multidim_exists_pos, multidim_assign_pos, multidim_bind_pos,
multidim_delete_pos, shaped_multidim_exists_pos, check_shaped_bounds,
make_not_enough_dimensions_error, allomorph_numeric_equal.

External callers that reference free functions via methods:: path: none
found — all callers use self.call_method_with_values() directly. The
fn/method multiset is byte-identical. No behavior change.
